+++ src/utils.h
@@ -1,7 +1,7 @@
 #include "Rcpp.h"
 #include "uv.h"
 
-inline Rcpp::CharacterVector asCharacterVector(std::string x) {
+inline Rcpp::CharacterVector asCharacterVector(const std::string& x) {
   return Rcpp::CharacterVector(Rf_mkCharCE(x.c_str(), CE_UTF8));
 }
 
@@ -10,4 +10,4 @@ inline Rcpp::CharacterVector asCharacterVector(std::string x) {
 uv_dirent_type_t get_dirent_type(
     const char* path, const uv_dirent_type_t& entry_type = UV_DIRENT_UNKNOWN);
 
-std::string path_tidy_(const std::string in);
+std::string path_tidy_(const std::string& in);