+++ tests/testthat/test-path.R
@@ -105,6 +105,17 @@ describe("path_tidy", {
     expect_equal(path_tidy("foo\\\\bar\\\\baz\\\\"), "foo/bar/baz")
   })
 
+  it("always appends windows root paths with /", {
+    expect_equal(path_tidy("C:"), "C:/")
+    expect_equal(path_tidy("c:"), "c:/")
+    expect_equal(path_tidy("X:"), "X:/")
+    expect_equal(path_tidy("x:"), "x:/")
+    expect_equal(path_tidy("c:/"), "c:/")
+    expect_equal(path_tidy("c://"), "c:/")
+    expect_equal(path_tidy("c:\\"), "c:/")
+    expect_equal(path_tidy("c:\\\\"), "c:/")
+  })
+
   it("passes NA along", {
     expect_equal(path_tidy(NA_character_), NA_character_)
     expect_equal(path_tidy(c("foo/bar", NA_character_)), c("foo/bar", NA_character_))
@@ -238,9 +249,9 @@ describe("path_norm", {
     expect_equal(path_norm("."), ".")
     expect_equal(path_norm(""), ".")
     expect_equal(path_norm("/"), "/")
-    expect_equal(path_norm("c:/"), "c:")
+    expect_equal(path_norm("c:/"), "c:/")
     expect_equal(path_norm("/../.././.."), "/")
-    expect_equal(path_norm("c:/../../.."), "c:")
+    expect_equal(path_norm("c:/../../.."), "c:/")
     expect_equal(path_norm("../.././.."), "../../..")
     expect_equal(path_norm("C:////a/b"), "C:/a/b")
     expect_equal(path_norm("//machine/share//a/b"), "//machine/share/a/b")